Move tasks to Buildvana.Sdk.Tasks namespace, base tasks to Buildvana.Sdk.Tasks.Base.
